Rename dockerapi module to docker
